build and push images

+++ gen_data_docker_image/python/build_image.py
@@ -1,17 +1,54 @@
+import rm_image_dockerhub
 import subprocess
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from pathlib import Path
+import requests
 
 # The directory containing the Dockerfiles
 directory_path = '/home/cc/Praxi-study/Praxi-Pipeline/gen_data_docker_image/python/dockerfiles'
 # Your Docker Hub username
-docker_hub_username = 'zongshun96'
+USERNAME = 'zongshun96'
+PASSWORD = ''  # Consider using a Personal Access Token for better security
+NAMESPACE = 'zongshun96'
+registry_url = 'https://registry.hub.docker.com/v2'
+index_url = 'https://index.docker.io/v2'
+dockerhub_front_url = 'https://hub.docker.com/v2'
+# /namespaces/{namespace}/repositories/{repository}/tags/{tag}
+def is_tag_pushed(repo_name, tag, token):
+    """Check if the tag is already pushed to Docker Hub."""
+    headers = {'Authorization': f'JWT {token}'}
+    response = requests.get(f"{dockerhub_front_url}/namespaces/{NAMESPACE}/repositories/{repo_name}/tags/{tag}", 
+                            headers=headers)
+    # response = requests.get(f"{dockerhub_front_url}/repositories/{USERNAME}/{repo_name}/tags/{tag}", 
+    #                         headers=headers)
 
-def build_push_remove_docker_image(dockerfile_path):
+    # ############## Directly querying the Docker Hub Backend #####################
+    # token = rm_image_dockerhub.get_backend_auth_token(repo_name)
+    # headers = {'Authorization': f'Bearer {token}'}
+    # response = requests.get(f"{index_url}/{USERNAME}/{repo_name}/manifests/{tag}", 
+    #                         headers=headers)
+    # #############################################################################
+    
+    return response.status_code == 200
+
+    # TOKEN=$(curl -s "https://auth.docker.io/token?service=registry.docker.io&scope=repository:zongshun96/python3_9_18-bookworm.prometheus-flask-exporter_v0_22_4:pull" | jq -r .token)
+    # curl -s -H "Authorization: Bearer $TOKEN" https://index.docker.io/v2/zongshun96/python3_9_18-bookworm.prometheus-flask-exporter_v0_22_4/manifests/TAG
+
+
+def build_push_remove_docker_image(dockerfile_path, token, idx):
     """Builds, pushes, and removes a Docker image from a specified Dockerfile."""
+
+    if idx % 20 == 0:
+        # Removes all build cache objects.
+        cleanup_build_cache()
+
     dockerfile_name = dockerfile_path.name
     # Create a tag using Docker Hub username and Dockerfile name, excluding 'Dockerfile.' prefix
-    tag = f"{docker_hub_username}/{dockerfile_name.replace('Dockerfile.', '')}"
+    tag = f"{USERNAME}/{dockerfile_name.replace('Dockerfile.', '')}"
+    if is_tag_pushed(dockerfile_name.replace('Dockerfile.', ''), "latest", token):
+        # print(f"Tag {tag} already exists on Docker Hub, skipping push")
+        return f"Tag {tag} already exists on Docker Hub, skipping push"
+
     build_cmd = ['docker', 'build', '-t', tag, '-f', str(dockerfile_path), str(dockerfile_path.parent)]
     push_cmd = ['docker', 'push', tag]
     remove_cmd = ['docker', 'rmi', tag]
@@ -19,18 +56,28 @@ def build_push_remove_docker_image(dockerfile_path):
     try:
         # Build the Docker image
         subprocess.run(build_cmd, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        print(f"Successfully built {tag}")
+        # print(f"Successfully built {tag}")
         
         # Push the Docker image
         subprocess.run(push_cmd, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        print(f"Successfully pushed {tag}")
+        # print(f"Successfully pushed {tag}")
         
         # Remove the local Docker image
         subprocess.run(remove_cmd, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        return f"Successfully removed {tag}"
+        # print(f"Successfully removed python3{tag}")
+
+        return f"Done {tag}"
     except subprocess.CalledProcessError as e:
         from pathlib import Path
-        Path('/home/cc/Praxi-study/Praxi-Pipeline/gen_data_docker_image/python/dockerfiles_failed/'+dockerfile_path.name).touch()
+        target_dockerfile_pathname = '/home/cc/Praxi-study/Praxi-Pipeline/gen_data_docker_image/python/dockerfiles_failed/'+dockerfile_path.name
+        if not Path(target_dockerfile_pathname).exists():
+            Path(target_dockerfile_pathname).touch()
+        with open(target_dockerfile_pathname, "a") as f:
+            f.write(e.stderr.decode("utf-8"))
+        with open(target_dockerfile_pathname, "a") as f:
+            f.write("=========================================\n")
+        with open(target_dockerfile_pathname, "a") as f:
+            f.write(e.stdout.decode("utf-8"))
         return f"Failed to build/push/remove {tag}: {e}"
 
 def find_dockerfiles(directory):
@@ -40,15 +87,28 @@ def find_dockerfiles(directory):
 
 def build_push_remove_images_in_parallel(dockerfiles):
     """Builds, pushes, and removes multiple Docker images in parallel from a list of Dockerfiles."""
-    with ThreadPoolExecutor() as executor:
-        future_to_dockerfile = {executor.submit(build_push_remove_docker_image, df): df for df in dockerfiles}
-        for future in as_completed(future_to_dockerfile):
+    token = rm_image_dockerhub.get_auth_token(USERNAME, PASSWORD)
+    with ThreadPoolExecutor(max_workers=128) as executor:
+        future_to_dockerfile = {executor.submit(build_push_remove_docker_image, df, token, idx): df for idx, df in enumerate(dockerfiles)}
+        for idx, future in enumerate(as_completed(future_to_dockerfile)):
             dockerfile = future_to_dockerfile[future]
             try:
                 result = future.result()
-                print(result)
+                print(f"{idx} {result}")
             except Exception as exc:
-                print(f'{dockerfile} generated an exception: {exc}')
+                print(f'{idx} {dockerfile} generated an exception: {exc}')
+
+
+def cleanup_build_cache():
+    """Removes all build cache objects."""
+    cleanup_cmd = ['docker', 'system', 'prune', '-f']
+    try:
+        result = subprocess.run(cleanup_cmd, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        print("Successfully cleaned up all build cache objects.")
+        # print(result.stdout)
+    except subprocess.CalledProcessError as e:
+        print(f"Failed to clean up build cache objects: {e.stderr.decode()}")
+
 
 if __name__ == "__main__":
     dockerfiles = find_dockerfiles(directory_path)
@@ -56,3 +116,8 @@ if __name__ == "__main__":
         build_push_remove_images_in_parallel(dockerfiles)
     else:
         print(f"No Dockerfiles found in {directory_path}.")
+
+    # cleanup_build_cache()
+
+
+
